requirements.txt
flask==3.0.0
paho-mqtt==1.6.1
numpy==1.26.2
scipy==1.11.4
pydantic==2.5.2
python-dotenv==1.0.0
shapely==2.0.2
waitress==3.0.0
opencv-python-headless==4.8.1.78
# ultralytics is optional, mock used if missing
# RPi.GPIO is hardware dependent
